+++ components/collector/src/source_collectors/robot_framework/tests.py
@@ -27,8 +27,9 @@ class RobotFrameworkTests(RobotFrameworkBaseClass):
             test_entities.extend(entities)
         return SourceMeasurement(value=str(nr_of_tests), total=str(total_nr_of_tests), entities=test_entities)
 
-    @staticmethod
+    @classmethod
     async def _parse_source_response(
+        cls,
         response: Response,
         test_results: list[str],
         all_test_results: list[str],
@@ -37,6 +38,7 @@ class RobotFrameworkTests(RobotFrameworkBaseClass):
         nr_of_tests, total_nr_of_tests, entities = 0, 0, Entities()
         tree = await parse_source_response_xml(response)
         stats = first(tree.findall("statistics/total/stat"), lambda stat: (stat.text or "").lower() == "all tests")
+        parent_map = cls.parent_map(tree)
         for test_result in all_test_results:
             total_nr_of_tests += int(stats.get(test_result, 0))
             if test_result not in test_results:
@@ -47,6 +49,13 @@ class RobotFrameworkTests(RobotFrameworkBaseClass):
                 for test in suite.findall(f"test/status[@status='{test_result.upper()}']/.."):
                     test_id = test.get("id", "")
                     test_name = test.get("name", "unknown")
-                    entity = Entity(key=test_id, test_name=test_name, suite_name=suite_name, test_result=test_result)
+                    suite_names = cls.parent_names(test, parent_map)
+                    entity = Entity(
+                        key=test_id,
+                        test_name=test_name,
+                        suite_name=suite_name,
+                        suite_names=suite_names,
+                        test_result=test_result,
+                    )
                     entities.append(entity)
         return nr_of_tests, total_nr_of_tests, entities