+++ components/collector/tests/source_collectors/robot_framework/test_tests.py
@@ -9,13 +9,23 @@ class RobotFrameworkTestReportTest(RobotFrameworkTestCase):
     METRIC_TYPE = "tests"
     LANDING_URL = "https://robot_framework"
 
+    def robot_entity(self, *, key: str, name: str, result: str, suite: str) -> dict[str, str]:
+        """Create a Robot Framework entity."""
+        return {
+            "key": key,
+            "test_name": name,
+            "test_result": result,
+            "suite_name": suite,
+            "suite_names": suite,
+        }
+
     async def test_tests(self):
         """Test that the number of tests is returned."""
         expected_entities = [
-            {"key": "s1-t1", "test_name": "Test 1", "test_result": "fail", "suite_name": "Suite 1"},
-            {"key": "s1-t2", "test_name": "Test 2", "test_result": "pass", "suite_name": "Suite 1"},
-            {"key": "s2-t1", "test_name": "Test 3", "test_result": "pass", "suite_name": "Suite 2"},
-            {"key": "s3-t1", "test_name": "Test 4", "test_result": "skip", "suite_name": "Suite 3"},
+            self.robot_entity(key="s1-t1", name="Test 1", result="fail", suite="Suite 1"),
+            self.robot_entity(key="s1-t2", name="Test 2", result="pass", suite="Suite 1"),
+            self.robot_entity(key="s2-t1", name="Test 3", result="pass", suite="Suite 2"),
+            self.robot_entity(key="s3-t1", name="Test 4", result="skip", suite="Suite 3"),
         ]
         for xml in self.ROBOT_FRAMEWORK_XMLS:
             with self.subTest(xml=xml):
@@ -31,7 +41,7 @@ class RobotFrameworkTestReportTest(RobotFrameworkTestCase):
     async def test_failed_tests(self):
         """Test that the number of failed tests is returned."""
         self.set_source_parameter("test_result", ["fail"])
-        expected_entities = [{"key": "s1-t1", "test_name": "Test 1", "test_result": "fail", "suite_name": "Suite 1"}]
+        expected_entities = [self.robot_entity(key="s1-t1", name="Test 1", result="fail", suite="Suite 1")]
         for xml in self.ROBOT_FRAMEWORK_XMLS:
             with self.subTest(xml=xml):
                 response = await self.collect(get_request_text=xml)
@@ -46,7 +56,7 @@ class RobotFrameworkTestReportTest(RobotFrameworkTestCase):
     async def test_skipped_tests(self):
         """Test that the number of skipped tests is returned."""
         self.set_source_parameter("test_result", ["skip"])
-        expected_entities = [{"key": "s3-t1", "test_name": "Test 4", "test_result": "skip", "suite_name": "Suite 3"}]
+        expected_entities = [self.robot_entity(key="s3-t1", name="Test 4", result="skip", suite="Suite 3")]
         for xml in self.ROBOT_FRAMEWORK_XMLS:
             with self.subTest(xml=xml):
                 response = await self.collect(get_request_text=xml)